backend/pyproject.toml
[project]
name = "incident-workbench"
version = "0.1.0"
description = "Incident analysis and clustering workbench"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "scikit-learn>=1.6.0",
    "numpy>=2.0.0",
    "python-docx>=1.1.0",
    "ollama>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pyinstaller>=6.0.0",
]

[tool.setuptools]
py-modules = []

[tool.setuptools.packages.find]
include = ["models*", "routers*", "services*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
